implement candidate calling for virus application
