+++ src/candidates/mod.rs
@@ -1,0 +1,2 @@
+pub mod hla;
+pub mod virus;
